Переделал структуру, удалил лишний файл
